+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/MethodHeadingNode.java
@@ -7,34 +7,10 @@ import org.sonar.plugins.delphi.antlr.DelphiLexer;
 import org.sonar.plugins.delphi.antlr.DelphiParser;
 import org.sonar.plugins.delphi.antlr.ast.node.FormalParameterNode.FormalParameter;
 import org.sonar.plugins.delphi.antlr.ast.visitors.DelphiParserVisitor;
+import org.sonar.plugins.delphi.symbol.declaration.MethodKind;
 import org.sonar.plugins.delphi.type.Type;
 
 public final class MethodHeadingNode extends DelphiNode {
-  public enum MethodKind {
-    CONSTRUCTOR(DelphiLexer.CONSTRUCTOR),
-    DESTRUCTOR(DelphiLexer.DESTRUCTOR),
-    FUNCTION(DelphiLexer.FUNCTION),
-    OPERATOR(DelphiLexer.OPERATOR),
-    PROCEDURE(DelphiLexer.PROCEDURE);
-
-    private final int tokenType;
-
-    MethodKind(int tokenType) {
-      this.tokenType = tokenType;
-    }
-
-    private static MethodKind fromTokenType(int tokenType) {
-      for (MethodKind kind : MethodKind.values()) {
-        if (kind.tokenType == tokenType) {
-          return kind;
-        }
-      }
-
-      throw new AssertionError(
-          "Unhandled MethodKind token type: " + DelphiParser.tokenNames[tokenType]);
-    }
-  }
-
   private String image;
   private Boolean isClassMethod;
   private String typeName;